Move broadcast tests to test_at_operator
