+++ metafunctions/tests/test_at_operator.py
@@ -1,3 +1,4 @@
+import unittest
 
 from metafunctions.util import node
 from metafunctions.tests.util import BaseTestCase
@@ -24,7 +25,30 @@ class TestUnit(BaseTestCase):
         c = (1, 2, 3) @ a
         self.assertEqual(str(c), '((1, 2, 3) | star(a))')
 
+    def test_upgrade_merge(self):
+        aabbcc = (a & b & c) @ (a&b&c)
+        self.assertEqual(aabbcc('_'), ('_aa', '_bb', '_cc'))
 
+    @unittest.skip('TODO')
+    def test_recursive_upgrade(self):
+        aabbcc = (a & b & c) @ star(a+b+c)
+        self.assertEqual(aabbcc('_'), '_aa_bb_cc')
+
+    def test_str_repr(self):
+        cmp = a @ (b&c)
+        self.assertEqual(str(cmp), '(a | star(b & c))')
+        self.assertEqual(repr(cmp),
+                f'FunctionChain({a!r}, SimpleFunction({cmp._functions[1]._function}))')
+
+    @unittest.skip('Map')
+    def test_loop(self):
+        cmp = (b & c & 'stoke') @ star(a)
+        self.assertEqual(cmp('_'), ('_ba', '_ca', 'stokea'))
+
+    @unittest.skip('Map')
+    def test_loop_with_non_meta(self):
+        cmp = (b & c & 'stoke') @ star(len)
+        self.assertEqual(cmp('_'), (2, 2, 5))
 
 @node
 def a(x):
@@ -32,3 +56,8 @@ def a(x):
 @node
 def b(x):
     return x + 'b'
+
+@node
+def c(x):
+    return x + 'c'
+
